Move credential utility functions to CognitoUtility to avoid circular dependency problems.
