+++ src/app/service/aws.service.ts
@@ -1,7 +1,6 @@
 import {Injectable} from "@angular/core";
 import {CognitoUtil, Callback} from "./cognito.service";
 import * as AWS from "aws-sdk/global";
-import * as CognitoIdentity from "aws-sdk/clients/cognitoidentity";
 
 /**
  * Created by Vladimir Budilov
@@ -14,40 +13,8 @@ export class AwsUtil {
     public static firstLogin: boolean = false;
     public static runningInit: boolean = false;
 
-    public cognitoCreds:AWS.CognitoIdentityCredentials;
-
-    constructor() {
+    constructor(public cognitoUtil: CognitoUtil) {
         AWS.config.region = CognitoUtil._REGION;
-    }
-
-    // AWS Stores Credentials in many ways, and with TypeScript this means that 
-    // getting the base credentials we authenticated with from the AWS globals gets really murky,
-    // having to get around both class extension and unions. Therefore, we're going to give
-    // developers direct access to the raw, unadulterated CognitoIdentityCredentials
-    // object at all times.
-    setCognitoCreds(creds:AWS.CognitoIdentityCredentials) {
-        this.cognitoCreds = creds;
-    }
-
-    getCognitoCreds(){
-        return this.cognitoCreds;
-    }
-
-    // This method takes in a raw jwtToken and uses the global AWS config options to build a
-    // CognitoIdentityCredentials object and store it for us. It also returns the object to the caller
-    // to avoid unnecessary calls to setCognitoCreds.
-
-    buildCognitoCreds(idTokenJwt:string) {
-        let url = 'cognito-idp.' + CognitoUtil._REGION.toLowerCase() + '.amazonaws.com/' + CognitoUtil._USER_POOL_ID;
-        let logins:CognitoIdentity.LoginsMap = {};
-        logins[url] = idTokenJwt;
-        let params = {
-            IdentityPoolId: CognitoUtil._IDENTITY_POOL_ID, /* required */
-            Logins: logins
-        };
-        let creds = new AWS.CognitoIdentityCredentials(params);
-        this.setCognitoCreds(creds);
-        return creds;
     }
 
     /**
@@ -117,7 +84,7 @@ export class AwsUtil {
     }
 
     addCognitoCredentials(idTokenJwt: string): void {
-        let creds = this.buildCognitoCreds(idTokenJwt);
+        let creds = this.cognitoUtil.buildCognitoCreds(idTokenJwt);
 
         AWS.config.credentials = creds;
